Rework the whole file storage layout
all files are now stored according to the following layout:

/path/to/mdr2/files/:artifact/:id/

where :artifact is similar to state but not quite. For each production
we produce a number of artifacts (xml files containing the structure,
obi projects, DTB exports from obi, iso files, etc). These are loosely
related to the state but the state reflects more at what point the
production is and what it is waiting for. :id is the id of the
production.

So essentially we have

- /path/to/mdr2/files/structured/:id/ where we keep the DTBook XML and
  the config file for obi
- /path/to/mdr2/files/recording/:id/ the obi project
- /path/to/mdr2/files/recorded/:id/ the export from obi (DAISY3 with
  wav)
- /path/to/mdr2/files/encoded/:id/ the encoded and downgraded DAISY3
- /path/to/mdr2/files/iso/:id/ the iso file(s)