Move validation to jing
Also moved schema validation into a separate namespace to that it can be
reused.

The old validation seemed to have a strange concept of valid. It claimed
that smil files were valid vubis export files. Jing seems to master this
quite well. The downside is that we have another obscure dependency.