Add a HTTP push interface for abacus
This required substantial changes to validation and massive reductions in
the ABACUS facing code.

There are now http endpoints i.e. api routes to deliver the files from
ABACUS.

The queues for metadata-updates and creation are gone. This is now done
directly upon the request. Also the cron jobs for polling for these
files are gone.