Emit total time according to standards
According to http://www.daisy.org/z3986/2005/Z3986-2005.html#Clock the
totalTime has to be displayed in the form of hh:mm:ss.ms. We store the
total time as a duration in milliseconds in the database. So we have to
convert this to a clock value when emitting the DTBook and the rdf
files. In other cases (such as for the catalog and for ABACUS) the total
time is expected in minutes.

Also for legacy reasons we have to put the dam number in the identifier
field when producing the rdf.

Fixes #33