Unpack the production to be repaired inside structured
We used to unpack productions that were to be repaired inside /tmp
before moving the relevant files to the proper place (inside
structured). However since /tmp is a small partition we now unpack the
production directly inside structured in a tmp directory which is
deleted after the relevant files have been moved.

Fixes #34