Retry iso generation
If the creation of the iso failed, wait for a moment and then try
again. There's a suspicion that the encoding process is finished but
the underlying storage hasn't managed to flush all the content.

So the hope is that after waiting a little while the iso creation
might work since the storage has caught up by now.