Revert "Upgrade immutant dependency"
This reverts commit de639b4d32bd9066851acbe4d7050e051287004e.

Upgrading to immutant 2.1.10 is nice but I can no longer install the
resulting war. I get

java.lang.UnsupportedClassVersionError: from/riddley/Util: Unsupported major.minor version 52.0

riddley is part of Potemkin (a dependency of immutant) which was
internalized in 2.1.6.

The madras production machine runs on Java 1.7 and Debian
7.11 (current release is 10), so there is not much hope for a newer
Java.

So for now the quickest is to just revert this change.