Drop the transaction FIXMEs where they made no sense
previously (with immutant) we had distributed transactions. These
could handle transactions across both a database and a message queue.
Now we do not have distributed transactions anymore, so we cannot have
transactions across a db and a message queue.

But mostly that is not a problem as if a problem occurs in the
database an exception is raised and the resulting non-local exit will
prevent the push to the message queue anyway.

So these transaction related FIXMEs cab safely be removed.