Use a "manual" transaction to rollback repair
The natural thing to do would of course be a transaction with
isolation level :none as we want others to see that the current
production is in state "repairing". But alas that level is not
supported by MySQL, so we have to resort to a "manual" transaction
using try/catch.